Fixed Jaggy Scrolling!!!!!!!
The scroll view does not quickly jump to a frame when snapping without
animating anymore. Instead of looking at the position of the cell, we
look at the velocity of the scroll and see if it goes in the positive
or negative direction and scroll accordingly. It practically does
exactly the same thing as before except now it always animates when it
changes states! Yay!!!!